devlink/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
mysqlclient==2.2.7
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.1
